requirements.txt
Jinja2==2.8
MarkupSafe==0.23
pkg-resources==0.0.0
PyYAML==3.12
